% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvInfer.R
\docType{methods}
\name{cnvInfer}
\alias{cnvInfer}
\title{Inferring CNV using different methods}
\usage{
cnvInfer(
  SeuratObj,
  group.by = "seurat_clusters",
  normal_groups = NULL,
  methods = c("infercnv"),
  nthreads = 8,
  outdir = "./",
  gene_order_file = NULL
)
}
\arguments{
\item{SeuratObj}{A seurat object including count matrix and cell cluster annotation.}

\item{group.by}{An integer or a character specifying the cell cluster column in the seurat meta.data.}

\item{normal_groups}{A vector specifying the groups of normal cells, which mush match the name shown in meta.data.}

\item{methods}{copykat, infercnv, numbat}

\item{nthreads}{An integer specifying the number of threads for running inferCNV}

\item{outdir}{path to output intermediate results.}

\item{gene_order_file}{path to the gene order file (required for infercnv); using geneorder_hg20 by default.}
}
\value{
A seurat object with prediction results added into meta.data.
}
\description{
Inferring CNV using different methods
}
\examples{

}
\author{
Wubing Zhang
}
