% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnvParser.R
\docType{methods}
\name{infercnvParser}
\alias{infercnvParser}
\title{Inferring CNV using different methods}
\usage{
infercnvParser(infercnv_obj, window.width = 1)
}
\arguments{
\item{infercnv_obj}{An Infercnv object.}

\item{window.width}{width of windows to group genes}
}
\value{
A data frame with inferred malignancy.
}
\description{
Inferring CNV using different methods
}
\examples{

}
\author{
Wubing Zhang
}
