% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnalyzer.R
\docType{methods}
\name{scAnalyzer}
\alias{scAnalyzer}
\title{Analysis pipeline for single cell RNA-seq data}
\usage{
scAnalyzer(
  obj,
  project = NULL,
  analyses = c("qc", "doublet", "norm", "pca", "clustering"),
  resolution = 0.1,
  outdir = NULL
)
}
\arguments{
\item{obj}{A matrix or data.frame of count table or a seurat object.}

\item{project}{A character, specifying the name of the project, which will be used in output file names.}

\item{analyses}{A character vector, specifying the type of analysis to perform, avaible analysis include qc,
doublet, norm, pca, clustering, findmarker.}

\item{resolution}{The resolution for clustering. 0.1 by default.}

\item{outdir}{The path to output figures.}
}
\value{
A list a seurat object and markers for all clusters
}
\description{
Analysis pipeline for single cell RNA-seq data
}
\details{
This pipeline takes count table as input, and performs data quality control, normalization, doublets removal (scDblFinder),
dimention reduction, cell clustering and identification of marker genes.
}
\examples{
library(scAnalyzer);
pbmc.data <- Read10X(data.dir = "../data/pbmc3k/filtered_gene_bc_matrices/hg19/")
res <- scAnalyzer(pbmc.data)
}
\author{
Wubing Zhang
}
