% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScAnnotator.R
\docType{methods}
\name{ScAnnotator}
\alias{ScAnnotator}
\title{Annotating cell types using SingleR}
\usage{
ScAnnotator(
  query.obj,
  clusters = "seurat_clusters",
  method = c("seurat"),
  ref.data = NULL,
  ref.ann = NULL,
  markers = NULL,
  plot = TRUE,
  outdir = "./"
)
}
\arguments{
\item{query.obj}{A seurat object for cell type annotation.}

\item{clusters}{A character specifying the cluster column in meta.data.}

\item{method}{A vector specifying the cell type annotation methods.
Should be one of singler, sclearn, scmap-cell, scmap-cluster, seurat}

\item{ref.data}{A matrix like object specifying the reference data or a Seurat object.}

\item{ref.ann}{A vector specifying the cell type annotation of the reference data.}

\item{markers}{A list with Positive (markers$Positive) and Negative markers (markers$Negative) of cell types.}

\item{plot}{A boolean indicating whether to save and visualize the prediction results.}

\item{outdir}{Path to the output directory.}
}
\value{
A seurat object with prediction results added into meta.data.
}
\description{
Annotating cell types using SingleR
}
\examples{

}
\author{
Wubing Zhang
}
