% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNormalize.R
\docType{methods}
\name{scNormalize}
\alias{scNormalize}
\title{Normalization of single-cell RNA-seq read counts}
\usage{
scNormalize(
  counts,
  norm.method = c("LogNormalize", "SCTransform", "downsample", "quminorm", "census"),
  mc.cores = 2,
  fraction = 0.85
)
}
\arguments{
\item{counts}{A matrix.}

\item{norm.method}{One of LogNormalize, SCTransform, downsample, quminorm, and census.}

\item{mc.cores}{An integer, specifying the number of cores for quminorm and downsample normalization.}

\item{fraction}{A numeric between 0 and 1, specifying the fraction of cells to be downsampling normalized.}
}
\value{
A normalized count matrix.
}
\description{
Normalization of single-cell RNA-seq read counts
}
\examples{

}
\author{
Wubing Zhang
}
