% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScType.R
\docType{methods}
\name{ScType}
\alias{ScType}
\title{Incorporate ScType predictions to Seurat object}
\usage{
ScType(query.obj, markers)
}
\arguments{
\item{query.obj}{A seurat object for cell type annotation.}

\item{markers}{A list with Positive (markers$Positive) and Negative markers (markers$Negative) of cell types.}
}
\value{
A seurat object with ScType results attached to meta.data.
}
\description{
Incorporate ScType predictions to Seurat object
}
