% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScType.R
\docType{methods}
\name{ScType}
\alias{ScType}
\title{ScType: calculate ScType scores and assign cell types}
\usage{
ScType(
  scRNAseqData,
  scaled = !0,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{scRNAseqData}{- input scRNA-seq matrix (rownames - genes, column names - cells),}

\item{gs}{- list of gene sets positively expressed in the cell type}

\item{gs2}{- list of gene sets that should not be expressed in the cell type (NULL if not applicable)}

\item{scale}{- indicates whether the matrix is scaled (TRUE by default)}
}
\value{
A data frame with ScType scores for each cell type.
}
\description{
GNU General Public License v3.0 (https://github.com/IanevskiAleksandr/sc-type/blob/master/LICENSE)
}
\author{
Aleksandr Ianevski <aleksandr.ianevski@helsinki.fi>, June 2021
}
